requirements.txt
# 기존 패키지
PyMuPDF>=1.23.0
pdfplumber>=0.10.0
tqdm>=4.66.0

# AWS SDK
boto3>=1.34.0
botocore>=1.34.0

# LangChain + AWS
langchain>=0.1.0
langchain-aws>=0.1.0
langchain-community>=0.0.10

# Vector Store
opensearch-py>=2.4.0

# 유틸리티
python-dotenv>=1.0.0
tenacity>=8.2.0
tiktoken>=0.5.0

# 데이터 처리
numpy>=1.24.0
pandas>=2.0.0

# 테스트
pytest>=7.4.0
pytest-asyncio>=0.21.0

# FastAPI 서버
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6

# 벡터 DB (로컬)
chromadb>=0.4.0
sentence-transformers>=2.2.0

# 문서 파싱
python-docx>=0.8.11
openpyxl>=3.1.0
olefile>=0.46

# KISTI ScienceON API
pycryptodome>=3.19.0
requests>=2.31.0
